Move beacon related code to reusable components
