+++ src/renderer/containers/AddBeaconNode/AddBeaconNode.tsx
@@ -7,5 +7,5 @@ export const AddBeaconNodeContainer: React.FunctionComponent = () => {
     return (
         <>
         </>
-    )
+    );
 };
