+++ src/renderer/containers/Onboard/OnboardContainer.tsx
@@ -15,7 +15,7 @@ import WithdrawalKey from "./WithdrawalKey/WithdrawalKey";
 import {WithdrawalKeyGenerateContainer} from "./WithdrawalKey/Generate/WithdrawalKeyGenerate";
 import {WithdrawalKeyVerifyContainer} from "./WithdrawalKey/Verify/WithdrawalKeyVerify";
 import {ConfigureContainer} from "./Configure/ConfigureContainer";
-import {ConfigureBeaconNode} from "./Configure/ConfigureBeaconNode";
+import {ConfigureBeaconNodeContainer} from "./Configure/ConfigureBeaconNode";
 
 interface IOnboardStep {
     step: string
@@ -37,7 +37,7 @@ export default class OnboardContainer extends Component<IProps, {}> {
         [OnBoardingRoutes.WITHDRAWAL_KEY_VALIDATE]: <WithdrawalKeyVerifyContainer history={this.props.history}/>,
         [OnBoardingRoutes.WITHDRAWAL_IMPORT]: <WithdrawalKeyImportContainer history={this.props.history} />,
         [OnBoardingRoutes.CONFIGURE]: <ConfigureContainer history={this.props.history}/>,
-        [OnBoardingRoutes.CONFIGURE_BEACON_NODE]: <ConfigureBeaconNode history={this.props.history}/>,
+        [OnBoardingRoutes.CONFIGURE_BEACON_NODE]: <ConfigureBeaconNodeContainer history={this.props.history}/>,
         [OnBoardingRoutes.CONFIGURE_DOCKER_PATH]: <ConfigureDockerPath history={this.props.history}/>,
         [OnBoardingRoutes.DEPOSIT_TX]: <DepositTxContainer history={this.props.history}/>,
         [OnBoardingRoutes.PASSWORD]: <CreatePasswordContainer history={this.props.history}/>,
